src/scripts/requirements.txt
boto3==1.21.16
botocore==1.24.16
fastapi==0.75.0
gcsfs==2022.2.0
geopy==2.2.0
h5py==3.2.1
matplotlib==3.2.0
numpy==1.20.3
pandas==1.3.4
pydantic==1.9.0
setuptools==58.0.4
tensorflow-cpu==2.8.0
tqdm==4.62.3
uvicorn==0.17.5
pillow==7.1.2
email-validator==1.1.2
PyJWT==1.7.1
python-decouple==3.3
google-cloud==0.34.0
google-cloud-bigquery==3.0.1